Cart Projector testes -woop- wip

+++ database/factories/ProductFactory.php
@@ -19,7 +19,7 @@ class ProductFactory extends Factory
             'name' => $this->faker->words('4', true),
             'description' => $this->faker->paragraph(2, true),
             'cost' => $cost,
-            'retail' => ($cost * config('shop.profit_margin')),
+            'retail' => (int)($cost * config('shop.profit_margin')),
             'active' => $this->faker->boolean(),
             'vat' => config('shop.vat'),
             'category_id' => Category::factory()->create(),
